# Fix method invalidations around `Base.convert`