--- a/src/webserver/Dynamic.jl
+++ b/src/webserver/Dynamic.jl
@@ -181,7 +181,7 @@ function send_notebook_changes!(🙋::ClientRequest; commentary::Any=nothing)
         if client.connected_notebook !== nothing && client.connected_notebook.notebook_id == 🙋.notebook.notebook_id
             current_dict = get(current_state_for_clients, client, :empty)
             patches = Firebasey.diff(current_dict, notebook_dict)
-            patches_as_dicts::Array{Dict} = patches
+            patches_as_dicts::Array{Dict} = Firebasey._convert(Array{Dict}, patches)
             current_state_for_clients[client] = deep_enough_copy(notebook_dict)
 
             # Make sure we do send a confirmation to the client who made the request, even without changes