--- a/src/webserver/Firebasey.jl
+++ b/src/webserver/Firebasey.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.9
+# v0.19.12
 
 using Markdown
 using InteractiveUtils
@@ -702,6 +702,18 @@ end
 md"### applypatch! AddPatch"
   ╠═╡ =#
 
+# ╔═╡ d7ea6052-9d9f-48e3-92fb-250afd69e417
+begin
+    _convert(::Type{Base.UUID}, s::String) = Base.UUID(s)
+    _convert(::Type{T}, a::AbstractArray) where {T<:Array} = _convert.(eltype(T), a)
+    _convert(x, y) = convert(x, y)
+
+    function _setproperty!(x, f::Symbol, v)
+        type = fieldtype(typeof(x), f)
+        return setfield!(x, f, _convert(type, v))
+    end
+end
+
 # ╔═╡ dd87ca7e-2de1-11eb-2ec3-d5721c32f192
 function applypatch!(value, patch::AddPatch)
 	if length(patch.path) == 0
@@ -721,7 +733,7 @@ function applypatch!(value, patch::AddPatch)
 			if strict_applypatch[]
 				@assert getproperty(subvalue, key) === nothing
 			end
-			setproperty!(subvalue, key, patch.value)
+			_setproperty!(subvalue, key, patch.value)
 		end
 	end
 	return value
@@ -758,7 +770,7 @@ function applypatch!(value, patch::ReplacePatch)
 			if strict_applypatch[]
 				@assert getproperty(subvalue, key) !== nothing
 			end
-			setproperty!(subvalue, key, patch.value)
+			_setproperty!(subvalue, key, patch.value)
 		end
 	end
 	return value
@@ -795,7 +807,7 @@ function applypatch!(value, patch::RemovePatch)
 			if strict_applypatch[]
 				@assert getproperty(subvalue, key) !== nothing
 			end
-			setproperty!(subvalue, key, nothing)
+			_setproperty!(subvalue, key, nothing)
 		end
 	end
 	return value
@@ -1220,6 +1232,7 @@ end
 # ╠═48a45941-2489-4666-b4e5-88d3f82e5145
 # ╠═752b2da3-ff24-4758-8843-186368069888
 # ╟─3e285076-1d97-4728-87cf-f71b22569e57
+# ╠═d7ea6052-9d9f-48e3-92fb-250afd69e417
 # ╠═dd87ca7e-2de1-11eb-2ec3-d5721c32f192
 # ╟─c3e4738f-4568-4910-a211-6a46a9d447ee
 # ╟─a11e4082-4ff4-4c1b-9c74-c8fa7dcceaa6