This PR fixes about 257 + 180 + 248 + 68 = 753 method invalidations.

Found via the following code on Julia 1.8.2:
```julia
julia> using SnoopCompileCore

julia> invs = @snoopr using Pluto;

julia> using SnoopCompile

julia> trees = invalidation_trees(invs);

julia> trees[end]
inserting convert(::Type{Base.UUID}, string::String) in Pluto at /home/rik/git/Pluto.jl/src/notebook/Cell.jl:77 invalidated:
   mt_backedges: 1: signature Tuple{typeof(convert), Type{Base.UUID}, Any} triggered MethodInstance for setindex!(::Dict{Base.UUID, Tuple{String, Union{Nothing, VersionNumber}}}, ::Any, ::Any) (2 children)
                 2: signature Tuple{typeof(convert), Type{Base.UUID}, Any} triggered MethodInstance for (::Base.var"#cvt1#1"{Tuple{String, String, Base.UUID}})(::Int64) (5 children)
                 3: signature Tuple{typeof(convert), Type{Base.UUID}, Any} triggered MethodInstance for (::Base.var"#cvt1#1"{Tuple{Base.UUID, String, String, VersionNumber}})(::Int64) (5 children)
                 4: signature Tuple{typeof(convert), Type{Base.UUID}, String} triggered MethodInstance for convert(::Type{Union{Nothing, Base.UUID}}, ::String) (27 children)
                 5: signature Tuple{typeof(convert), Type{Base.UUID}, Any} triggered MethodInstance for setindex!(::Dict{Base.UUID, Tuple{String, Union{Nothing, VersionNumber}}}, ::Tuple{String, Any}, ::Any) (180 children)
                 6: signature Tuple{typeof(convert), Type{Base.UUID}, Any} triggered MethodInstance for setindex!(::Dict{String, Base.UUID}, ::Any, ::String) (257 children)
```
and

```julia
julia> trees[end-1]
inserting convert(::Type{<:Dict}, patch::Pluto.Firebasey.ReplacePatch) in Pluto.Firebasey at /home/rik/git/PlutoFork.jl/src/webserver/Firebasey.jl:231 invalidated:
   mt_backedges:  1: signature Tuple{typeof(convert), Type{Dict{String, Union{String, Vector{String}}}}, Any} triggered MethodInstance for setindex!(::Vector{Dict{String, Union{String, Vector{String}}}}, ::Any, ::Int64) (0 children)
                  2: signature Tuple{typeof(convert), Type{Dict{String, String}}, Any} triggered MethodInstance for setindex!(::Vector{Dict{String, String}}, ::Any, ::Int64) (0 children)
                  3: signature Tuple{typeof(convert), Type{Dict{String, Any}}, Any} triggered MethodInstance for Artifacts._artifact_str(::Module, ::String, ::SubString{String}, ::String, ::Dict{String, Any}, ::Base.SHA1, ::Base.BinaryPlatforms.Platform, ::Any) (0 children)
                  4: signature Tuple{typeof(convert), Union{Type{Dict{Char, Any}}, Type{REPL.LineEdit.CompletionProvider}, Type{REPL.LineEdit.HistoryProvider}}, Any} triggered MethodInstance for setproperty!(::REPL.LineEdit.HistoryPrompt, ::Symbol, ::Any) (1 children)
                  5: signature Tuple{typeof(convert), Type{Dict{String, String}}, Any} triggered MethodInstance for setindex!(::Dict{String, Dict{String, String}}, ::Any, ::String) (3 children)
                  6: signature Tuple{typeof(convert), Type{Dict{Symbol, REPL.LineEdit.Prompt}}, Any} triggered MethodInstance for REPL.REPLHistoryProvider(::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any) (4 children)
                  7: signature Tuple{typeof(convert), Type{Dict{Char, Any}}, Any} triggered MethodInstance for REPL.LineEdit.Prompt(::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any, ::Any) (5 children)
                  8: signature Tuple{typeof(convert), Type{Dict{Symbol, Any}}, Any} triggered MethodInstance for setindex!(::IdDict{Function, Dict{Symbol, Any}}, ::Any, ::Any) (8 children)
                  9: signature Tuple{typeof(convert), Type{Dict{String, Any}}, Any} triggered MethodInstance for setindex!(::Dict{Base.BinaryPlatforms.AbstractPlatform, Dict{String, Any}}, ::Any, ::Base.BinaryPlatforms.Platform) (68 children)
                 10: signature Tuple{typeof(convert), Type{Dict{String, Union{String, Vector{String}}}}, Any} triggered MethodInstance for setindex!(::Dict{String, Dict{String, Union{String, Vector{String}}}}, ::Any, ::String) (248 children)
```