@rikhuijzer 2022-09-30
In e321baa, I've worked around the failing front end tests by redirecting `setproperty!` to `_setproperty!` and defining some additional conversions on a more local level. The benefit of this [layer of indirection](https://medium.com/@nmckinnonblog/indirection-fba1857630e2) is that it avoids having to define global overrides and hence it avoids method invalidations. In turn, this should reduce TTFX.