@rikhuijzer 2022-10-01
It's hard to benchmark the gains, but allocations appear to be reduced which is as expected. I've taken the benchmarks below from the CI output logs for Ubuntu Julia 1.8. Ignore the time since CPUs may have been different.

Allocations for the whole Pluto test suite went from 13.8 GiB to 13.2 GiB which means a 4% reduction.

# main

```
────────────────────────────────────────────────────────────────────────
                                              Time          Allocations  
                                        ───────────────   ───────────────
             Total measured:                  979s            13.8GiB    

 Section                        ncalls     time    %tot     alloc    %tot
 ────────────────────────────────────────────────────────────────────────
 import Pluto                        1    2.75s    0.3%    232MiB    1.6%
 compiletimes.jl                     1    43.0s    4.4%   1.81GiB   13.1%
   PlutoRunner.run_expression        1    1.89s    0.2%   69.9MiB    0.5%
   SessionActions.open               1    25.7s    2.6%    774MiB    5.5%
   Pluto.run                         1    4.86s    0.5%    284MiB    2.0%
 Events.jl                           1    3.72s    0.4%    237MiB    1.7%
 WorkspaceManager.jl                 1     100s   10.3%    686MiB    4.9%
 packages/Basic.jl                   1     207s   21.4%   1.25GiB    9.1%
 Bonds.jl                            1    49.1s    5.1%    266MiB    1.9%
 RichOutput.jl                       1    52.5s    5.4%    771MiB    5.5%
 React.jl                            1     200s   20.6%   3.75GiB   27.2%
 Dynamic.jl                          1    36.5s    3.8%    685MiB    4.8%
 MacroAnalysis.jl                    1     169s   17.4%   1.09GiB    7.9%
 Logging.jl                          1    17.0s    1.7%   61.5MiB    0.4%
 webserver.jl                        1    49.9s    5.1%    689MiB    4.9%
 Notebook.jl                         1    12.1s    1.2%   0.99GiB    7.2%
 Configuration.jl                    1    8.17s    0.8%    441MiB    3.1%
 ReloadFromFile.jl                   1    8.50s    0.9%    144MiB    1.0%
 packages/PkgCompat.jl               1    681ms    0.1%   63.5MiB    0.4%
 ExpressionExplorer.jl               1    6.53s    0.7%    430MiB    3.0%
 MethodSignatures.jl                 1    186ms    0.0%   17.3MiB    0.1%
 MoreAnalysis.jl                     1    1.21s    0.1%    116MiB    0.8%
 Analysis.jl                         1    148ms    0.0%   7.02MiB    0.0%
 webserver_utils.jl                  1   81.0ms    0.0%   1.17MiB    0.0%
 data structures.jl                  1    807ms    0.1%   66.8MiB    0.5%
 DependencyCache.jl                  1    1.01s    0.1%   81.3MiB    0.6%
 Throttled.jl                        1    843ms    0.1%   2.08MiB    0.0%
 cell_disabling.jl                   1    499ms    0.1%   32.9MiB    0.2%
 ────────────────────────────────────────────────────────────────────────
```

# This PR

```
 ────────────────────────────────────────────────────────────────────────
                                              Time          Allocations  
                                        ───────────────   ───────────────
             Total measured:                  909s            13.2GiB    

 Section                        ncalls     time    %tot     alloc    %tot
 ────────────────────────────────────────────────────────────────────────
 import Pluto                        1    2.30s    0.3%    216MiB    1.6%
 compiletimes.jl                     1    37.6s    4.2%   1.67GiB   12.7%
   PlutoRunner.run_expression        1    1.58s    0.2%   67.6MiB    0.5%
   SessionActions.open               1    21.4s    2.4%    677MiB    5.0%
   Pluto.run                         1    4.61s    0.5%    279MiB    2.1%
 Events.jl                           1    3.96s    0.4%    251MiB    1.9%
 WorkspaceManager.jl                 1    89.9s   10.0%    675MiB    5.0%
 packages/Basic.jl                   1     197s   21.8%   1.25GiB    9.5%
 Bonds.jl                            1    46.3s    5.1%    265MiB    2.0%
 RichOutput.jl                       1    49.7s    5.5%    767MiB    5.7%
 React.jl                            1     185s   20.5%   3.34GiB   25.4%
 Dynamic.jl                          1    35.3s    3.9%    727MiB    5.4%
 MacroAnalysis.jl                    1     153s   17.0%   1.08GiB    8.2%
 Logging.jl                          1    15.9s    1.8%   55.7MiB    0.4%
 webserver.jl                        1    47.6s    5.3%    680MiB    5.1%
 Notebook.jl                         1    11.2s    1.2%   0.96GiB    7.3%
 Configuration.jl                    1    8.00s    0.9%    436MiB    3.2%
 ReloadFromFile.jl                   1    8.52s    0.9%    144MiB    1.1%
 packages/PkgCompat.jl               1    618ms    0.1%   63.5MiB    0.5%
 ExpressionExplorer.jl               1    6.43s    0.7%    404MiB    3.0%
 MethodSignatures.jl                 1    169ms    0.0%   16.9MiB    0.1%
 MoreAnalysis.jl                     1    1.10s    0.1%    105MiB    0.8%
 Analysis.jl                         1    140ms    0.0%   7.03MiB    0.1%
 webserver_utils.jl                  1   77.3ms    0.0%   1.13MiB    0.0%
 data structures.jl                  1    772ms    0.1%   62.9MiB    0.5%
 DependencyCache.jl                  1    415ms    0.0%   32.5MiB    0.2%
 Throttled.jl                        1    838ms    0.1%   2.08MiB    0.0%
 cell_disabling.jl                   1    472ms    0.1%   31.5MiB    0.2%
 ────────────────────────────────────────────────────────────────────────
```