@rikhuijzer 2022-10-03
I'll merge since the changes can easily be reversed at any point in the future if that would become neccessary.

Thanks for the review @Pangoraw 💚